+++ app/src/main/java/com/example/cocktaildb/screen/myrecipe/MyRecipeFragment.kt
@@ -12,12 +12,9 @@ import com.example.cocktaildb.data.model.Cocktail
 import com.example.cocktaildb.data.repository.CocktailRepository
 import com.example.cocktaildb.data.repository.source.local.CocktailLocalDataSource
 import com.example.cocktaildb.databinding.FragmentMyRecipeBinding
-import com.example.cocktaildb.databinding.ItemCocktailBinding
 import com.example.cocktaildb.utils.base.BaseFragment
 
-/**
- * Fragment displaying user's created recipes
- */
+
 class MyRecipeFragment : BaseFragment<FragmentMyRecipeBinding>(), MyRecipeContract.View {
 
     private lateinit var presenter: MyRecipePresenter
@@ -99,52 +96,6 @@ class MyRecipeFragment : BaseFragment<FragmentMyRecipeBinding>(), MyRecipeContra
 
     override fun displayError(message: String) {
         Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-    }
-
-    /**
-     * Adapter for the recipes RecyclerView
-     */
-    private inner class RecipeAdapter : RecyclerView.Adapter<RecipeAdapter.RecipeViewHolder>() {
-
-        private val recipes: MutableList<Cocktail> = mutableListOf()
-
-        fun setRecipes(recipes: Collection<Cocktail>) {
-            this.recipes.clear()
-            this.recipes.addAll(recipes)
-            notifyDataSetChanged() // Consider using more specific change notifications
-        }
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecipeViewHolder {
-            val binding = ItemCocktailBinding.inflate(
-                LayoutInflater.from(parent.context), parent, false
-            )
-            return RecipeViewHolder(binding)
-        }
-
-        override fun onBindViewHolder(holder: RecipeViewHolder, position: Int) {
-            holder.bind(recipes[position])
-        }
-
-        override fun getItemCount(): Int = recipes.size
-
-        inner class RecipeViewHolder(private val binding: ItemCocktailBinding) :
-            RecyclerView.ViewHolder(binding.root) {
-
-            fun bind(cocktail: Cocktail) {
-                binding.tvCocktailName.text = cocktail.strDrink
-                binding.tvCocktailCategory.text = cocktail.strCategory ?: getString(R.string.custom_recipe)
-
-                // Load the cocktail image using our custom ImageLoader
-                com.example.cocktaildb.utils.ImageLoader.loadImage(
-                    url = cocktail.strDrinkThumb,
-                    imageView = binding.ivCocktail,
-                    placeholderResId = R.drawable.ic_launcher_background
-                )
-
-                // Set rating for my recipes
-                binding.tvRating.text = getString(R.string.my_recipe)
-            }
-        }
     }
 
     companion object {
